Mark `dest` as optional in SEP-0006 withdraw
If an anchor is requiring interactive customer information, it may be possible to omit `dest` in the initial withdraw API call.  The reason being that the anchor can collect the dest information interactively while they collect the customer information.

I would propose that anchors that want to declare that dest is required should do so via the Info endpoint.