Allow 404 /transaction responses (#359)
* Allow 404 /transaction responses

Make the interactive flow easier to implement for anchors by allowing 404 responses for incomplete transactions. This means the anchor will not need to create a transaction entry in its database as soon as `/deposit` is hit. Instead, it can wait for successful completion of the interactive flow, and the wallet will handle that correctly.

Also add a note for anchors to test their interactive forms on mobile.

* Improve wording