Add Deprecated status to SEP Process
Create a deprecated status and explain how to move a SEP into it.  Its important that we remove any cruft from our SEPs to ensure people have a clear understanding of the state of the ecosystem, and to prevent people from implementing SEPs that won't be compatible with anyone else.