doc: make README engaging (#465)
Change the beginning of the `README` to include the Stellar logo, and Stellar's mission.

See it fully rendered at:
https://github.com/leighmcculloch/stellar--protocol/blob/sparkle/README.md

Why:

At GitHub Universe one of the talks I attended talked about how creating an inviting and engaging README makes a huge difference to communicating to developers who are coming across the project. The [dev.to README] was given as an example of a great README that goes a long way to engage poeple who visit the project. I think there's probably a lot of things we can do to make our README more alive to someone visiting it, and adding our logo and including our mission is just one small thing.

The same change has been merged on the `stellar/go` repository in stellar/go#1948.

[dev.to README]: https://github.com/thepracticaldev/dev.to/blob/master/README.md