SEP-0010: Add best practices about JWTs and URIs (#468)
Make really clear that the issuer is a URI and not a URL, and add notes about where to find best practices about JWTs.

It's really easy to see `URI` and just think `URL` even for us who are rather intimate with the SEP. An JWTs issuer can be a URL, but it also doesn't need to be, it can be a unique URI containing any information, such as the key used for signing. It is application specific and this clarification loosens the definition without prescribing how an application should populate the issuer field.

JWT best practices are important when implementing systems that use JWTs for authentication. It's out of scope for this SEP to capture those best practices, also because they are evolving, but we can link to the current best practices that the IETF are drafting.