SEP-0011: Add headings to values sections
### What
Add headings to the Values section of the document with each type and unique encoding situation getting a heading.

### Why
I find myself frequently needing to link to sections in this document, specifically the section about how to encode `Asset` as a string. Without headings it is not possible to link directly to the discussion about the `Asset` type and I can only instead link people to the section about values and hope that they read through to find what I'm referencing. It makes it easier to share concepts and behaviors defined in this SEP if they can be easily linked to.