SEP-0018: Distill down to focus on namespace syntax and definition (#544)
### What
Distill SEP-18 to focus solely on namespace syntax and definition, and
leave parsing and interpretation of, parsing, and rendering of a set of
data entries to applications and/or a future SEP.

### Why
The main gains we get from SEP-18 immediately is for
other SEPs to define data entries that do not collide with other
applications by placing them in a namespace that other applications know
should be avoided. The SEP also attempts to define how to parse the data
entries into trees, but this can be implemented in many different ways
and has unique challenges that deserve being the focus of their own SEP.
It should be quite easy to build another SEP on-top of SEP-18.