[SEP24] Add explicit amount field to api
To make a nice UX many wallets wish to collect the amount of deposit/withdrawal ahead of time to show fees before going through the whole interactive process.  We had originally loosely suggested providing it to the interactive url, but that doesn't help calculate fees ahead of time, and wasn't well documented.  Here we add it to the params expected in the initial POST calls.