SEP-0030: Add proposal for recoverysigner multi-party key management (#553)
### What
Add a proposal for recoverysigner, multi-party key management of Stellar accounts where no server managing keys has control of the account, and supporting users:
1. Recovering access to Stellar accounts for which they may have lost keys.
2. Gaining access to Stellar accounts that another user intends to share with them.

### Why
I have been experimenting with this proposal in implementation first over in the [exp/services/recoverysigner] package which implements most of what is documented here. I think I've experimented enough with it to warrant documenting.

### Expectations
I expect this proposal to remain in draft state for sometime while the finer details of the SEP are ironed out. I'd like to merge the SEP without perfecting it so that I can iterate on it, and keep it and the experimental implementation in line so that the experimental implementation has a well defined specification as it evolves, and so that as the specification changes it is captured clearly outside of code.

[exp/services/recoverysigner]: https://github.com/stellar/go/blob/master/exp/services/recoverysigner