SEP-0030: Be less prescriptive about order of pagination
### What
Remove the statement about the order of the results that are returned by
the `GET /accounts` endpoint.

### Why
Ordering by when the accounts were registered is not of significance
since no fields about the registration time are included in the response
and there is nothing that we can imagine right now that a client would
do with that information. We should limit the SEPs requirements to those
that matter.