SEP-0010: Add info about why network_passphrase is included
### What
Add information about why the network_passphrase field is provided in
the response.

### Why
On a quick read of the docs readers have assumed the presence of the
network_passphrase field is for the server to tell the client which
passphrase to use. While there may be valid uses of the field in that
way the intention for why the field was added was to be a debugging
mechanism so that it is immediately clear if a client or server as
misconfigured. The reason this is helpful in that situation is because
the network passphrase affects the signature of the transaction but is
not stored in the transaction, so if a client or server are configured
incorrectly the error message from the server when verifying the
transaction will simply be that the signature is invalid.