SEP-0010: Note behavior for non-existent accounts in Abstract
### What
Add notes to the Abstract section about the optional behavior with Stellar accounts that do not yet exist.

### Why
This behavior is already detailed in the section that describes the behavior of the server when handling the Token endpoint. This is not new behavior, but not mentioning the behavior in the abstract has led to some confusion because a reader who only reads the abstract will think that an account must exist to be compatible with SEP-10.

The version bump as a patch number is due to no functional changes to the proposal.