SEP-0030: Remove deprecated fields and endpoint
### What
Remove the /sign endpoint and the `signer` field in the common account
response.

### Why
The endpoint and field were marked as deprecated. They make less sense
in the new version of SEP-30 where the client selects a signer and in
the interest of keeping the API as simple as possible and encourage
clients to choose a signer these should be removed.