SEP-0030: Remove deprecated fields and endpoint (#629)
### What
Remove the /sign endpoint and the `signer` field in the common account response.

### Why
The endpoint and field were marked as deprecated in v0.3.0. They make less sense in the new version of SEP-30 where the client selects a signer and in the interest of keeping the API as simple as possible and encourage clients to choose a signer these should be removed. Breaking changes are possible because this SEP is still in draft and the breaking changes were separated out from v0.3.0 to make the proposed change clearer for implementations based on the draft to navigate the change.