ALL: Remove use of the terms whitelist/blacklist
### What
Remove use of the terms whitelist and blacklist, replacing them with
other descriptions of what functionality is intended or with other terms
allowlist and denylist.

### Why
The terms whitelist and blacklist make many people feel unwelcome and
can trigger racial stereotyping. A similar change was made on the Go
repository which prompted me to check if we used these terms in our
documentation or code.
golang/go@608cdcaede1e7133dc994b5e8894272c2dce744b

The terms whitelist and blacklist are not better at communicating intent
than simply describing the behavior we're looking for or than using
another term. This seems like an easy change to make with no cost that
improves the general language of our documentation. An example of how
this generally improves the documentation is this sentence where we
state:

>...test them with your wallet to be sure they work before allowing them to be used with your wallet.

Which is much clearer and easier to read than the original:

>...test them with your wallet to be sure they work before whitelisting them.

This changes previously accepted documents. I don't think that is
particularly an issue given than the meaning of the documents remains
the same.