SEP-????: Add a proposal for user-friendly asset addresses
### What
Add a proposal for user-friendly asset addresses.

### Why
There does not seem to be a consistent user-friendly approach to
referencing or resolving unique assets in products that are built on
Stellar. Each wallet or application I use has a different approach and
it is an unpleasant experience to find assets using their issuer
addresses.