SEP-0032: Add proposal for user-friendly asset addresses and process for verifying them (#639)
### What
Add a proposal for user-friendly asset addresses that includes a process for verifying them.

### Why
There does not seem to be a consistent user-friendly approach to referencing or resolving unique assets in products that are built on Stellar. Each wallet or application I use has a different approach and it is an unpleasant experience to find assets using their issuer addresses.