SEP-0005: Replace uses of the term account with key
### What
Replace uses of the term account with key in all cases where the SEP is
likely to be used for a Stellar key that isn't necessarily a Stellar
account.

### Why
A Stellar account is a key which has led to some of our documentation
confusingly referencing an account when in fact the SEP is intended to
work with any key. This is an important distinction for accounts that
use multi-sig. SEP-5 is appropriate for multi-sig accounts and for
generating keys that will be used as signers for an account where that
signer is just a key and not an account.