SEP-0030: Remove the added_at field from signers
### What
Remove the `added_at` field from signers.

### Why
The field adds a burden on the implementer to keep track of when a
signing key is generated and stored but without any substantial value to
the client.

Servers that use shared signing keys are likely to store and configure
their signing keys through systems that don't store metadata like time
alongside them. For example if a key is simply stored as a secret in
their deployment system it may not have a time. In our reference
implementation this has resulted in us returning a zero time in some
cases.

We originally included the field on the assumption that a client may
like to know when a key became available. In practice however that
information is irrelevant to a client. A client, especially used by a
consumer, isn't going to display the time the signer became available to
the user. The presence of the time also doesn't change the recommended
action which is to update to the first key in the list.

We can always revisit re-adding the field if a real need and use case
arise but with our existing use cases it doesn't look like the field
provides any utility and is one more requirement for implementers,
albeit small.