SEP-0030: Add security section about external authentication methods (#655)
### What
Add security section about external authentication methods and highlighting that a SEP-30 implementation chooses what external authentication methods to support and clients choose which to use when registering accounts.

### Why
The intention is that the external authentication methods are flexible and the methods use make sense for the use case, both functionally, and from the perspective of security and risk. The document already takes this position but this section clarifies that intent.