Draft of CAP-0034 (change closetime semantics)
First draft of CAP-0034, intended to address
stellar-protocol issue #622, by changing the semantics of
the selection of the closetime when combining candidates in
the SCP nomination protocol.

The current nomination protocol produces an unsigned
("basic") form of union of the candidates which uses the
maximum closetime of all candidates as the closetime to
nominate.

The protocol proposed by this CAP would produce a signed
StellarValue, whose transaction set and upgrades are chosen
in the same way as in the current protocol, but whose
closetime is precisely the one in the same StellarValue as
the best transaction set.

(The change to closetime semantics is sufficient by itself
to address stellar-protocol issue #622; the change to make
the nominated StellarValue signed is an opportunity that we
take which the closetime semantics change provides.  The
signature is on the <txSet, closetime> pair, so by taking
the exact closetime from the same StellarValue as the chosen
transaction set, we allow ourselves to use the signature
from the node that proposed that StellarValue.)