Second draft of CAP-0034 (`closeTime` change)
Many revisions to the first draft of CAP-0034,
"Preserve Transaction-Set/Close-Time Affinity During
Nomination".

The presentation has been expanded in many ways, trimmed
in others, and organized around the shape of a table of
semantics changes from earlier code through the recent
PR 2608 optimizations to the behavior proposed in this CAP,
upon which we attempt to build a quasi-inductive "proof of
desirability" of the proposed changes.