Integrate CAP-0034 second draft review comments (#691)
- Make the effect of the proposal on
`minTime`/`txTOO_EARLY` validation explicit; formerly it
had mentioned only `maxTime`/`txTOO_LATE`.  Point out that
the change makes the nomination and ballot protocols use
the exact same `closeTime`-related conditions to trim and
validate transactions as the externalization (ledger close)
code does.

- Incorporate further code review comments to correct and
clarify several pieces of SCP-related wording and
presentation.

- Correct "maximal `closeTime` injection" sections to
explain the attack and the degree of amelioration more
accurately.

- Clean up confusing (and in at least one place, just
wrong) section #4 of "Detailed argument in favor".

- Remove misleading references to transaction sets or
`StellarValue`s "winning"; specify that it is being
selected for the ballot protocol, or is being externalized
(whichever is correct in the specific contexts).

- Remove misleading references to "best" transaction sets
in favor of simply "composited".