SEP-0010: Revert home domain verification and relax manage data operation constraints (v2.1.0) (#740)
### What
Revert home domain verification and relax manage data operation constraints allowing additional manage data operations as long as their source account is the server account.

### Why
There's been some ambiguity about the intended contents of home domain and in the interest of encouraging compatibility between servers and clients we need to revert the home domain verification that the SEP and SDKs are requiring on that field.

The relaxing of the manage data operations so that we the option of adding a new field for hostname verification. We haven't decided if we should add a new field or clear up the misunderstanding of the home domain field. Allowing additional manage data operations with the server account as the source account will allow future releases to add additional fields of data without breaking older clients. Requiring the source account to be the server account ensures that if there is a need in the future to add new fields that the client should confirm and sign for, that older clients won't blindly sign those challenge transactions.

This change is backwards compatible with v2.0.0.

Co-authored-by: Jake Urban <10968980+JakeUrban@users.noreply.github.com>