SEP-0030: Make compatible with CAP-33 (#744)
### What
Allow SEP-30 servers to sign transactions that contain source accounts other than the authenticated account, at the discretion of the SEP-30 server operator.

### Why
SEP-30 is not compatible with CAP-33 because SEP-30 requires servers to reject transactions that contain other accounts as source accounts. For accounts that have signers sponsored by a sponsoring account as supported by CAP-33, changing a sponsor requires operations with the sponsoring account as a source account. This means it is not possible to change a sponsored signer via a SEP-30 server.

The limitation within SEP-30 is to address the confused deputy problem. The SEP-30 server is signing a transaction, potentially with a shared key, and it must ensure that it only signs a transaction for a user that relate to accounts that the user is authenticated for.

In the case that a SEP-30 server operator is also the operator of a solution that is utilizing CAP-33, it is highly unlikely that the operator needs the sponsoring account to be registered with the SEP-30 server. In that case it is simple for the operator to ensure that the signing keys used by the SEP-30 server are never signers of the sponsoring account and therefore it doesn't matter if the SEP-30 server signs transactions containing the sponsoring account.

There may be other features of Stellar that make it easier to sign for specific operations on transactions and not an entire transaction, and if they ever eventuate this change would become redundant and could be deprecated. At the moment this appears to be the simplest path forward.