SEP-0010: Clarify the
### What
Clarify that the first manage data operation is the operation containing the client account, nonce, and home domain.

### Why
The first manage data operation has always been the first operation. This is stated explicitly further down in the SEP when the challenge transaction structure is described. All SDKs I have reviewed also follow this pattern. The abstract is somewhat ambiguous though because it states that there is a single manage data operation with those values without stating it is the first.

### Note
This change is only improving the language and is not adding or changing the functionality of the SEP.