SEP-0030: Change title to Account Recovery
### What
Change the title to `Account Recovery` and remove mention of key management in title.

### Why
The SEPs purpose is to provide a multi-party approach to recovering Stellar accounts, either accounts a user created or owns, or accounts that another user has created and shared with the user. It is not technically key management because it is an interoperability layer that sits in the middle of an approach to key management.

The title `Recoverysigner` has also not been very useful. It was added to the beginning of the title to provide a single word keyword that could be used to reference SEP-30, but it turns out the term is odd, and so isn't being used to reference SEP-30.