SEP-0030: Add clearer instruction that other authentication types are supported and recommended formats for their use to improve interoperabilty
### What
Add clearer instruction that other authentication types are supported and recommended formats for their use.

### Why
SEP-30 has always stated that implementers may choose any authentication methods, and there are no limitations within SEP-30 that require an authentication method to be synchronous and immediate. However, there have been comments from people who have considered the SEP that suggest that is not widely known even amongst people who are familiar with the technical details of the SEP.

Also to improve interoperabilty. Without standard and common approaches to communicating about a variety of identities, SEP-30 server APIs will diverge significantly. There are other issues regarding interoperability in that external authentication is not defined in anyway within this SEP, however this change does not address that.