SEP-0010: Add back home domain verification with clarity (v3.0.0) (#746)
### What
Add back the home domain verification, but with greater clarity about exactly what the home domain is.

### Why
In v2.1.0 the home domain verification that was added in v2.0.0 was removed because the exact contents of the home domain field could be interpreted in at least two different ways that led to inconsistent implementations. The home domain is still present in the SEP-10 challenge transaction and this change provides clarity to what it should contain and how clients should verify it. This change represents the intended understanding of the home domain field. See v2.0.0 and v2.1.0 for prior context.

### Notes
Servers should use this change to verify that they are placing the correct value for the home domain into the challenge transaction. Once clients implement v3.0.0 they will reject challenge transactions from servers who are providing home domain values that are incorrect.

Co-authored-by: Sergey Nebolsin <nebolsin@gmail.com>
Co-authored-by: Jake Urban <jake@stellar.org>
Co-authored-by: Alex Cordeiro <accordeiro@users.noreply.github.com>