SEP-0030: Be unopinionated about number of servers required by protocol
### What
Be unopinionated about number of servers required by protocol and do not exclude use of the protocol with a single server, but include note that the single server configuration means control of the account.

### Why
The protocol specifies an interface a client and server can use to facilitate account recovery through transaction signing and authentication with providers outside of the Stellar network. It supports a configuration where a client uses multiple servers so that no individual server has control of the account on its own, but that is one use case and the protocol can support being used with a client and server configuration where there is only one server. There is little reason to limit the SEP to only multiple servers as it is up to the client how many servers they integrate with and utilize and are valid use cases where a client may wish to use a single server.