SEP-0030: Add clearer instruction that other authentication types are supported (#765)
### What
Add clearer instruction that other authentication types are supported and recommended formats for their use.

### Why
SEP-30 has always stated that implementers may choose any authentication methods, and there are no limitations within SEP-30 that require an authentication method to be synchronous and immediate. However this has not been clear as the SEP only discusses phone and email authentication in any real detail.

This change does not yet improve interoperability when other authentication methods are used because the formats for those authentication methods are not yet defined. I'll follow up with that in separate PRs.