CAP-0036: Add proposal for ClawbackClaimableBalanceOp as a sibling to CAP-35 (#797)
### What
Add proposal for ClawbackClaimableBalanceOp as a sibling to CAP-35 that proposes a method similar to CAP-35's clawback of an asset within an account that claws back an asset within a claimable balance.

### Why
The issue of claimable balances keeps occurring in relation to CAP-35 and this seems like the most explicit and simplest approach to also allowing issuers to clawback amounts in claimable balances. This should be considered in conjunction with CAP-35 but does not need to be tied to it and is otherwise independent.