SEP-8: reformat (#813)
What

This PR beautifies and re-formats SEP-8 without changing the functionality. It also makes things clear by adding a section "authorization flags" in order to explain how to identify a regulated asset in the first place. Finally, account setup is removed from Overview because it's not necessary. However, this PR does not remove the Account Setup section as its goal is to improves readability.

Why

A portion of the content and format in SEP-8 is outdated as it was created more than 2 years ago.