SEP-8: add application/json as request content type #816
What

Support JSON encoded request.

Why

We accept the application/json content type in some other SEPs, e.g. SEP-10 and SEP-30. The viewpoint from some people outside SDF that led to this pattern on SEP-10 was that it makes it easier to integrate with the APIs in some cases.