SEP-8: replace account setup with transaction composition (#817)
What

This PR replaces Account Setup with Transaction Composition.

Why

SEP-8 was introduced before CAP-18 was introduced. The Account Setup section in the current SEP-8 has a few drawbacks as mentioned in CAP-18.

1. The owner of an account cannot unilaterally utilize assets unrelated to the issuer without the signature of that issuer
2. If multiple issuers want to become signers on a single account, then many signatures are potentially required for simple operations
3. Issuers cannot easily rotate keys, because their keys are on every account holding their asset
We are seeing a few active developments on regulated assets in the ecosystem, and we want to make sure the associated SEP is up-to-date and is easy to use without introducing breaking changes.