SEP-8: add meaning to zero timeout in the case of pending #818
What

Make 0 timeout mean that the issuer could not determine the timeout needed.

Why

Issuers sometimes need to take extra steps before approving a transaction. Those steps may require manual actions.

In order to avoid breaking changes, we can use 0 to indicate the case where issuers couldn't determine the timeout at the time of returning the response. If we don't add this note, it can be misleading as 0 time might as well mean that the wallet/user can re-submit the transaction immediately.