SEP-0008: Add additional optional action params for the action required response (#758)
### What
Add functionality to the action_required status response and following behaviors providing:
- Support for a POST request.
- Support for early exit on the POST request.
- Support for passing parameters to the action URL using the GET or POST.

### Why
Clients often have some of the information that an asset issuer needs to collect from the user. This allows a client to supply that information upfront so that the user does not need to re-enter the information again.

This change assumes some minimal coordination between a wallet and server since a wallet will probably only provide information that it knows the server will accept. This coordination is truly minimal and a lack of coordination should not hamper an integration because the provision of the information is always optional.

The addition of the POST method allows the passing of sensitive information outside of the URL.

This change is backwards compatible for any client and server that support a GET action_url.

Co-authored-by: Leigh McCulloch <leigh@stellar.org>