SEP-12 callback endpoint (#805)
Wallets need a way to check a user's KYC status even if the user isn't available to sign a SEP-10 request. This change adds a POST /customer endpoint that allows wallets to provide a callback URL. Anchors will POST whenever the status changes (in the same format as responses to GET /customer) until the user is accepted or rejected.