SEP-0010: Fix the flow defined by the token endpoint since v3.1
### What
Make clear in the token endpoint flow that there are multiple operations in the challenge transaction and that the verification steps are focused on the first operation.

### Why
In SEP-10 v3.1 we added a new operation that is included but the token endpoint flow was not updated to reflect that this was the case. The token endpoint flow talks about the transaction having a single operation and verifying that operation which is confusing.

This change makes no functional changes and updates the token endpoint section to align with the rest of the document.

Co-authored-by: Jake Urban <jake@stellar.org>