SEP-0008: Add discussion about when to use a revised status and when not to (#898)
### What
Add discussion about when to use a revised status and when not to. Specifically distinguish that a revised transaction should never change the intent or semantics of the operations the user provided.

### Why
There was some conversation about this recently within SDF and we realize it is not clear in SEP-8 when to use reject or when to use revised. In general revised shouldn't alter the intent. If the user wants to perform an operation and it can succeed if the server adds operations, that's the perfect use case for revised. If the user wants to perform an operation that is not compliant no matter what operations can be added, it should just reject the transaction rather than try and modify the transaction.