SEP-0033: Set version and make active (#891)
Make SEP-33 active and set the version to v1.

The SEP is being used in several products in the ecosystem, namely stellar.expert, the Lobstr.co wallet, and I think some other places. There are several implementations, and the number of implementations is growing (#890). These are all indicators that this proposal/protocol is more than just a draft at this point. It's actively in use and the status should reflect that.